requirements.txt
# StoryRag v2.0 项目依赖
# 使用国内镜像安装

# PyTorch CUDA版本 (GPU加速)
torch>=2.7.0
torchaudio>=2.7.0
torchvision>=0.22.0

# LangChain框架
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# 向量嵌入和检索
sentence-transformers>=2.7.0
transformers>=4.39.0
accelerate>=0.28.0
safetensors>=0.4.0

# 文档处理
PyPDF2>=3.0.0
pypdf>=4.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0

# 数据处理和科学计算
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# 环境变量管理
python-dotenv>=1.0.0

# 日志和配置
loguru>=0.7.0
pydantic>=2.5.0

# 数据库和存储 (sqlite3和pathlib是Python标准库，无需安装)

# HTTP请求
httpx>=0.27.0
requests>=2.31.0

# JSON和数据序列化
orjson>=3.9.0

# Web框架
fastapi>=0.104.0
uvicorn>=0.24.0

# 向量数据库
chromadb>=0.4.0
faiss-cpu>=1.7.0
pinecone-client>=2.2.0

# 文本处理 (jieba将在需要时单独安装)
# spacy>=3.7.0

# PDF parsing backend (fallback to pdfminer.six for scanned/complex layouts)
pdfminer.six>=20221105
pdf2image>=1.16.3
pytesseract>=0.3.10
opencv-python>=4.9.0

# 可视化和调试 (matplotlib将在需要时单独安装)
# seaborn>=0.13.0

# 开发和测试工具 (将在需要时单独安装)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0